APC-Hackbench: Hold wake_lock before starting
Some runs could take an astonishingly high number of time. I think it
was going into deep sleep mode and holding the wake_lock cured it.

Also increase the sleep to 10 seconds so that the device doesn't go too
hot to quickly if run too many tests in a row.

TODO: Add a new PARAM for number of iterations.

Signed-off-by: Qais Yousef <qyousef@layalina.io>