Split out required tools into their own folder
Found a neat way to do that without having to ask the user to copy tools
around into nodes and without us having to code every tool required
inside a job.

The major difficulty that needed solving was in moving the tool from
master to the node. See 'Copy Tools' stage for how this is done.

Also make sure to clean up properly in the 'post always' step.

Signed-off-by: Qais Yousef <qyousef@layalina.io>